chore(proxy-http): Move utilities out of h1 module
Various utility functions are defined in the h1 module that are not strictly
HTTP/1-specific.

In preparation for additional reorganization of the http module, this change
moves utility functions:
