+++ linkerd/proxy/http/src/lib.rs
@@ -79,6 +79,12 @@ impl HasH2Reason for ::h2::Error {
     }
 }
 
+impl HasH2Reason for ::hyper::Error {
+    fn h2_reason(&self) -> Option<h2::Reason> {
+        (self as &(dyn std::error::Error + 'static)).h2_reason()
+    }
+}
+
 /// Returns an Authority from the value of `header`.
 pub fn authority_from_header<B, K>(req: &http::Request<B>, header: K) -> Option<Authority>
 where
@@ -86,4 +92,46 @@ where
 {
     let v = req.headers().get(header)?;
     v.to_str().ok()?.parse().ok()
+}
+
+fn set_authority(uri: &mut uri::Uri, auth: uri::Authority) {
+    let mut parts = uri::Parts::from(std::mem::take(uri));
+
+    parts.authority = Some(auth);
+
+    // If this was an origin-form target (path only),
+    // then we can't *only* set the authority, as that's
+    // an illegal target (such as `example.com/docs`).
+    //
+    // But don't set a scheme if this was authority-form (CONNECT),
+    // since that would change its meaning (like `https://example.com`).
+    if parts.path_and_query.is_some() {
+        parts.scheme = Some(http::uri::Scheme::HTTP);
+    }
+
+    let new = http::uri::Uri::from_parts(parts).expect("absolute uri");
+
+    *uri = new;
+}
+
+fn strip_connection_headers(headers: &mut http::HeaderMap) {
+    if let Some(val) = headers.remove(header::CONNECTION) {
+        if let Ok(conn_header) = val.to_str() {
+            // A `Connection` header may have a comma-separated list of
+            // names of other headers that are meant for only this specific
+            // connection.
+            //
+            // Iterate these names and remove them as headers.
+            for name in conn_header.split(',') {
+                let name = name.trim();
+                headers.remove(name);
+            }
+        }
+    }
+
+    // Additionally, strip these "connection-level" headers always, since
+    // they are otherwise illegal if upgraded to HTTP2.
+    headers.remove(header::UPGRADE);
+    headers.remove("proxy-connection");
+    headers.remove("keep-alive");
 }
